requirements.txt
DateTime==4.3
requests==2.27.1
html5lib==1.1
python-dotenv==0.19.2